// src/store/subAdmin/subAdmin.enum.ts
enum SubAdminTypeEnum {
  GET_SUB_ADMIN_PENDING = "GET_SUB_ADMIN_PENDING",
  GET_SUB_ADMIN_SUCCESS = "GET_SUB_ADMIN_SUCCESS",
  GET_SUB_ADMIN_FAILED = "GET_SUB_ADMIN_FAILED",

  GET_SUB_ADMIN_BY_ID_PENDING = "GET_SUB_ADMIN_BY_ID_PENDING",
  GET_SUB_ADMIN_BY_ID_SUCCESS = "GET_SUB_ADMIN_BY_ID_SUCCESS",
  GET_SUB_ADMIN_BY_ID_FAILED = "GET_SUB_ADMIN_BY_ID_FAILED",

  GET_ALL_ROLES_PENDING = "GET_ALL_ROLES_PENDING",
  GET_ALL_ROLES_SUCCESS = "GET_ALL_ROLES_SUCCESS",
  GET_ALL_ROLES_FAILED = "GET_ALL_ROLES_FAILED ",

  ADD_SUB_ADMIN_PENDING = "ADD_SUB_ADMIN_PENDING",
  ADD_SUB_ADMIN_SUCCESS = "ADD_SUB_ADMIN_SUCCESS",
  ADD_SUB_ADMIN_FAILED = "ADD_SUB_ADMIN_FAILED",

  UPDATE_SUB_ADMIN_PENDING = "UPDATE_SUB_ADMIN_PENDING",
  UPDATE_SUB_ADMIN_SUCCESS = "UPDATE_SUB_ADMIN_SUCCESS",
  UPDATE_SUB_ADMIN_FAILED = "UPDATE_SUB_ADMIN_FAILED",

  DELETE_SUB_ADMIN_PENDING = "DELETE_SUB_ADMIN_PENDING",
  DELETE_SUB_ADMIN_SUCCESS = "DELETE_SUB_ADMIN_SUCCESS",
  DELTE_SUB_ADMIN_FAILED = "DELTE_SUB_ADMIN_FAILED",

  UPDATE_SUB_ADMIN_STATUS_PENDING = "UPDATE_SUB_ADMIN_STATUS_PENDING",
  UPDATE_SUB_ADMIN_STATUS_SUCCESS = "UPDATE_SUB_ADMIN_STATUS_SUCCESS",
  UPDATE_SUB_ADMIN_STATUS_FAILED = "UPDATE_SUB_ADMIN_STATUS_FAILED",
}
export default SubAdminTypeEnum;